Add async operation end strategy for kotlin coroutines flow
