--- instrumentation/kotlinx-coroutines/javaagent/gradle.properties
@@ -1,1 +1,0 @@
-kotlin.stdlib.default.dependency=false
