+++ settings.gradle.kts
@@ -353,7 +353,9 @@ include(":instrumentation:kafka:kafka-clients:kafka-clients-0.11:testing")
 include(":instrumentation:kafka:kafka-clients:kafka-clients-2.6:library")
 include(":instrumentation:kafka:kafka-clients:kafka-clients-common:library")
 include(":instrumentation:kafka:kafka-streams-0.11:javaagent")
-include(":instrumentation:kotlinx-coroutines:javaagent")
+include(":instrumentation:kotlinx-coroutines:kotlinx-coroutines-1.0:javaagent")
+include(":instrumentation:kotlinx-coroutines:kotlinx-coroutines-flow-1.3:javaagent")
+include(":instrumentation:kotlinx-coroutines:kotlinx-coroutines-flow-1.3:javaagent-kotlin")
 include(":instrumentation:ktor:ktor-1.0:library")
 include(":instrumentation:ktor:ktor-2.0:javaagent")
 include(":instrumentation:ktor:ktor-2.0:library")
